+++ CHANGELOG.md
@@ -18,19 +18,22 @@ All user visible changes to this project will be documented in this file. This p
         - Swipe to reply gesture. ([#188], [#134])
         - Drafts. ([#173], [#102])
     - Chats tab:
-        - Chat muting/unmuting. ([#172], [#63])
+        - Chat muting/unmuting. ([#251], [#172], [#63])
         - Favorite chats. ([#218], [#209])
         - Searching. ([#206], [#205])
     - Home page:
         - Quick status changing menu. ([#204], [#203])
     - Media panel:
         - Participants redialing. ([#241], [#233])
+    - Contacts tab:
+        - Favorite contacts. ([#237], [#223])
 
 ### Changed
 
 - UI:
     - Home page:
         - Redesigned chats tab. ([#142])
+        - Redesigned introduction and logout modals. ([#249])
         - Redesigned menu tab. ([#244], [#243])
     - Media panel:
         - Video resizing when dragged. ([#191], [#190])
@@ -43,6 +46,10 @@ All user visible changes to this project will be documented in this file. This p
     - Profile page:
         - Redesigned profile page. ([#244], [#243])
     - Redesigned desktop context menu. ([#245])
+    - Auth page:
+        - Redesigned login modal. ([#249])
+    - User page:
+        - Redesigned user page. ([#254], [#252])
 
 ### Fixed
 
@@ -81,12 +88,18 @@ All user visible changes to this project will be documented in this file. This p
 [#218]: /../../pull/218
 [#221]: /../../pull/221
 [#232]: /../../issues/232
+[#223]: /../../issues/223
 [#233]: /../../issues/233
+[#237]: /../../pull/237
 [#241]: /../../pull/241
 [#243]: /../../issues/243
 [#244]: /../../pull/244
 [#245]: /../../pull/245
 [#246]: /../../pull/246
+[#249]: /../../pull/249
+[#251]: /../../pull/251
+[#252]: /../../issues/252
+[#254]: /../../pull/254